+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [Issue #557](https://github.com/influxdb/influxdb/issues/557). Group by time(1y) doesn't work while time(365d) works
 - [Issue #460](https://github.com/influxdb/influxdb/issues/460). Don't start automatically after installing
 - [Issue #529](https://github.com/influxdb/influxdb/issues/529). Don't run influxdb as root
+- [Issue #547](https://github.com/influxdb/influxdb/issues/547). Add difference function (Thanks, @mboelstra)
 
 ## v0.6.5 [2014-05-19]